Add new recipes and link to home page
Added the two new recipes per the requirements of the project and have
linked the two new recipes to the homepage so they are represented by
hyperlinks that once clicked will be opened in new tabs. Also, added
content to both of the recipes so that they resemble what a recipe
should look like.